example_4/api/requirements/development.txt
selenium
pytest
flask_restx
pytest-flask
webdriver_manager
pytest-httpserver
pytest-docker
pytest-mock